MD Settings: Add a non-pattern way to set origin permissions for Site Details.
Currently, setting origin-scoped permissions in "Site Details" calls
HostContentSettingMap::SetContentSettingCustomScope(). This is technically
supposed to only be used for setting ContentSettingPatterns, which can cross
over multiple origins, if the pattern allows for it.

To fix, implement a new method SiteSettingsHandler::HandleSetOriginPermissions,
which uses HostContentSettingMap::SetNarrowestContentSetting(), which only
handles origin-scoped permissions. This is more appropriate for Site Details,
which shows permissions for a specific origin.

Bug: 656758
Test: Navigate to https://permission.site. Open the Page Info bubble, then click
"Site settings". When the "Site Details" page opens, change a few of the
permissions listed there away from the original setting, and verify that
a) The drop-down doesn't immediately switch back to the original setting.
b) Switching back to the tab with https://permission.site and opening the Page
Info bubble shows the change that was just made accurately.
c) Refreshing the "Site Details" page still shows the new setting in the
drop-downs.
Also try clicking the "Clear and Reset" button underneath all the permissions,
and verify that it sets all the permissions back to their default values, and
that this is reflected in the Page Info bubble as well as "Site Details".

+++ content/browser/devtools/protocol/input_handler.h
@@ -57,8 +57,8 @@ class InputHandler : public DevToolsDomainHandler,
 
   void DispatchMouseEvent(
       const std::string& type,
-      double x,
-      double y,
+      int x,
+      int y,
       Maybe<int> modifiers,
       Maybe<double> timestamp,
       Maybe<std::string> button,
@@ -78,20 +78,20 @@ class InputHandler : public DevToolsDomainHandler,
   Response SetIgnoreInputEvents(bool ignore) override;
 
   void SynthesizePinchGesture(
-      double x,
-      double y,
+      int x,
+      int y,
       double scale_factor,
       Maybe<int> relative_speed,
       Maybe<std::string> gesture_source_type,
       std::unique_ptr<SynthesizePinchGestureCallback> callback) override;
 
   void SynthesizeScrollGesture(
-      double x,
-      double y,
-      Maybe<double> x_distance,
-      Maybe<double> y_distance,
-      Maybe<double> x_overscroll,
-      Maybe<double> y_overscroll,
+      int x,
+      int y,
+      Maybe<int> x_distance,
+      Maybe<int> y_distance,
+      Maybe<int> x_overscroll,
+      Maybe<int> y_overscroll,
       Maybe<bool> prevent_fling,
       Maybe<int> speed,
       Maybe<std::string> gesture_source_type,
@@ -101,8 +101,8 @@ class InputHandler : public DevToolsDomainHandler,
       std::unique_ptr<SynthesizeScrollGestureCallback> callback) override;
 
   void SynthesizeTapGesture(
-      double x,
-      double y,
+      int x,
+      int y,
       Maybe<int> duration,
       Maybe<int> tap_count,
       Maybe<std::string> gesture_source_type,
